_skels/python-django-skel/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "python-django-skel"
version = "1.0.0"
description = "Python/Django skeleton project (DRF + JWT + wrapper-shared backend stack)"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
dependencies = [
    "Django>=5.1,<6.0",
    "djangorestframework>=3.15.2,<4.0",
    # Wrapper-shared JWT bearer auth. simplejwt is the de-facto DRF
    # auth backend and ships HS256 + custom claims out of the box.
    "djangorestframework-simplejwt>=5.3.1,<6.0",
    "python-dotenv>=1.0.1",
    "gunicorn>=23.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.4,<9.0",
    "pytest-django>=4.9.0,<5.0",
    "pytest-cov>=6.0.0,<7.0",
    "ruff>=0.8.0,<1.0",
]

[tool.setuptools]
packages = ["myproject", "app", "app.migrations"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "C4", "SIM"]
ignore = ["E501"]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "myproject.settings"
python_files = ["test_*.py", "*_test.py"]
addopts = "-v --tb=short -p no:cacheprovider"
filterwarnings = [
    # simplejwt emits deprecation warnings on Django 5.x for internal
    # helpers that are irrelevant to our usage.
    "ignore::DeprecationWarning:rest_framework_simplejwt",
]
